Use adwaita theme and implement tracing
The adwaita theme provides a standard method of working with the user
interface. This also adds a range of useful tools, including those to
use the default desktop theming.

Tracing provides instrumentation to the application, making it easy to
debug and understand what is going on and where time is being taken.
This is particularly helpful when working with async functions, which
will definitely be coming down the pipeline.