Load images on separate threads
This is a huge perceived performance uplift. By loading the images on a
separate threadpool, there is no longer any lag to opening the window
and scrolling is generally completely fine. There are definitely still
some bugs, particularly in the scroll when going super fast, however
when stepping through the pictures there should be fairly minimal
slowdown.